Introduce `stepper` binary
This takes three values:
 * A current value
 * A direction to move (one of up, down, top, or bottom)
 * A config file containing a list of values to step up or down on

It then loads the config file to get a list of possible values to step
up or down on, and walks "up" or "down" from the "current value".

For example, with the inputs:

current-value: 2
direction: up
config-file contents:
```
1
3
5
7
```

would return "3", the next value "up" from "2" in that list.

"direction" values "top" and "bottom" return the last and first values
in the list found in the config file.